am62ax: udma: Fix static analysis KW issue rule INVARIANT_CONDITION.GEN
INVARIANT_CONDITION.GEN
Expression 'isCacheCoherent!= ( (Bool)1)' used in the condition
always yields the same result
Expression '( (uint16_t)65535) !=req.type' used in the
condition always yields the same result

Fixes: SITSW-7048

+++ source/drivers/udma/v0/udma_ring_lcdma.c
@@ -427,7 +427,7 @@ void Udma_lcdmaRingaccMemOps(void *pVirtAddr, uint32_t size, uint32_t opsType)
 {
     uint32_t    isCacheCoherent = Udma_isCacheCoherent();
 
-    if(isCacheCoherent != TRUE)
+    if(isCacheCoherent != (uint32_t)TRUE)
     {
         if(CSL_LCDMA_RINGACC_MEM_OPS_TYPE_WR == opsType)
         {
